feat: add initial database migration and update entity configurations
- Added initial EF Core migration (InitialCreate) for PostgreSQL

- Refined entity properties, constraints, and relationships in AppDbContext

- Added new enum and resolved enum mapping issues

- Implemented SaveChangesAsync timestamp interceptor for UpdatedAt properties

+++ src/ShopService/ShopService.Domain/Entities/OrderItem.cs
@@ -13,8 +13,8 @@ namespace ShopService.Domain.Entities
         public Guid? ProductId { get; set; }
         public string ProductName { get; set; } = string.Empty;
         public int Quantity { get; set; }
-        public double UnitPrice { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal TotalPrice { get; set; }
 
         public Order? Order { get; set; }
         public Product? Product { get; set; }
